% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/privacyrightsorg.R
\name{breaches.description.wordcloud}
\alias{breaches.description.wordcloud}
\title{breaches.description.wordcloud}
\usage{
breaches.description.wordcloud(breaches = privacyrightsorg::breaches,
  random.seed = 1234, min.freq = 50)
}
\arguments{
\item{breaches}{data.frame containing privacyrights.org breach database}

\item{random.seed}{seed for reproducibility default 1234}

\item{min.freq}{minimum word frequency default 50}
}
\value{
plot
}
\description{
Create wordlcoud with most commonly occurring breach description words
}

