% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{breaches}
\alias{breaches}
\title{privacyrights.org breach database}
\format{An object of class \code{"data.frame"};}
\source{
\href{https://www.privacyrights.org/data-breach}{Chronology of Data Breaches Security Breaches 2005 - Present}
}
\usage{
data('breaches')
}
\description{
Data from the privacyrights.org breach database in data.frame
format
}
\examples{
data('breaches')
summary(breaches)
}
\references{
\href{https://www.privacyrights.org/data-breach-FAQ}{Data breach FAQ}
}
\keyword{datasets}

