% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/privacyrightsorg.R
\name{breaches.records.timeseries.plot}
\alias{breaches.records.timeseries.plot}
\title{breaches.records.timeseries.plot}
\usage{
breaches.records.timeseries.plot(breaches = privacyrightsorg::breaches)
}
\arguments{
\item{breaches}{data.frame containing privacyrights.org breach database}
}
\value{
plot
}
\description{
Plot count of records breached as a monthly time series
}

