% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/privacyrightsorg.R
\name{parse.page}
\alias{parse.page}
\title{parse.page}
\usage{
parse.page(pg)
}
\arguments{
\item{pg}{html text to be parsed}
}
\value{
data.frame containing breach incident information (date, name, location,
entity.type, breach.type, records.compromised, description, info.links,
info.source)
}
\description{
Parse downloaded html pages from the privacyrights.org breach database
}

