% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/privacyrightsorg.R
\name{refresh.data}
\alias{refresh.data}
\title{refresh.data}
\usage{
refresh.data(data.file = file.path(system.file(c("inst", "extdata"), package =
  "privacyrightsorg"), "privacyrightsorg-html.rdata"), polite.pause = 1.5,
  download.data = TRUE)
}
\arguments{
\item{data.file}{rdata file where we cache copies of all downloaded html pages}

\item{polite.pause}{number of seconds to wait between http requests default 1.5}

\item{download.data}{whether or not to download all html data AND rebuild the db
(TRUE) or to just rebuild db with existing cached html data (FALSE). Default is FALSE}
}
\value{
list containing html content of each page in the breach database
}
\description{
Download all available html pages from the privacyrights.org breach database
at \url{http://privacyrights.org/data-breach}. This method can be called whenever
you wish to update your local copy of the database
}

